Fix the decimal precision issue. Update the rest of the codebase use quantity_to_receive and gwx_to_transfer attributes. Addition code refactor.

+++ db/migrate/20190628074903_add_quantity_to_receive_to_top_up_transactions.rb
@@ -1,7 +1,7 @@
 class AddQuantityToReceiveToTopUpTransactions < ActiveRecord::Migration[5.2]
 
   def change
-    add_column :top_up_transactions, :quantity_to_receive, :decimal, precision: 10, scale: 8
+    add_column :top_up_transactions, :quantity_to_receive, :decimal, precision: 16, scale: 8
   end
   
 end
